+++ lib/quickly.dart
@@ -6,4 +6,5 @@ export 'src/flutter/extensions/extensions.dart';
 export 'src/flutter/mixins/mixins.dart';
 export 'src/flutter/navigation/fx_animation.dart';
 export 'src/flutter/navigation/fx_navigation.dart';
-export 'src/flutter/responsive.dart';
+export 'src/flutter/responsive/fx_grid.dart';
+export 'src/flutter/responsive/fx_responsive.dart';